requirements.txt
-e .
GitPython==3.1.30
iniconfig==2.0.0
platformdirs==2.6.2
pre-commit==3.0.2
py==1.11.0
pytest==7.2.1
PyYAML==6.0
pydantic==1.10.4
python-slugify==8.0.0
typing-extensions==4.4.0
virtualenv==20.17.1
yamllint==1.29.0
yq==3.1.0
